Refactor intelligent reviewer into modular components
