+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -486,6 +486,9 @@ Code quality, organization, and maintainability over strict line count limits
 16. ✅ **CRIT-017**: OSRS AI Agent Test Suite (900 → 160 lines)
     - Split monolithic tests into game setup, agent, scenario, and validation modules
     - Added orchestrator for targeted execution
+17. ✅ **MAJOR-003**: Intelligent Reviewer Modularization
+    - File: `src/ai_ml/intelligent_reviewer.py`
+    - Split into core, AI analysis, reporting, and config modules with lightweight orchestrator
 
 ### Impact:
 - **Total lines reduced**: 14,962 → 3,466 lines (76.8% reduction)
